refactor(types): Minor Type Improvements for the `ThirdParty` Recipe (#946)
* refactor(types): Make `userContext` in `TypeProvider` methods optional

It is safe to pass `undefined` for `userContext` in all of these
cases. Therefore, to improve the DX for developers not leveraging
`UserContext` for these methods, the argument should be optional.

* refactor(types): Clarify `redirectURIQueryParams` Type

This argument expects an object of strings. This should be
clarified to developers so that they don't mistakenly pass,
for instance, an instance of `URLSearchParams`.

* chore: Changelog updates

---------

Co-authored-by: Rishabh Poddar <rishabh.poddar@gmail.com>